+++ src/tts/gather-text.ts
@@ -10,7 +10,7 @@ export function getReadableText(orig: string): string {
         .replace(/\\c\[[^\]]*\]/g, '')
         .replace(/\\s\[[^\]]*\]/g, '')
         .replace(/%/g, ' percent')
-        .replace(/\+/g, ' and ')
+        .replace(/\+/g, ' plus')
 
     const imgMatches: string[] | null = text.match(/\\i\[[^\]]*\]/g)
     for (let img of imgMatches ?? []) {
@@ -147,6 +147,8 @@ export class TextGather {
         this.initGather()
     }
 
+    recivedItemRecord: Record<sc.ItemID, number> = {}
+
     initPoststart() {
         const self = this
         let lastArea: string | undefined
@@ -182,6 +184,26 @@ export class TextGather {
                 }
             })()
         )
+
+        sc.Model.addObserver<sc.PlayerModel>(
+            sc.model.player,
+            new (class {
+                modelChanged(model: sc.Model, msg: sc.PLAYER_MSG, data: sc.PLAYER_MSG_ITEM_OBTAINED_DATA) {
+                    if (Opts.tts && model == sc.model.player && !sc.model.isCutscene() && msg == sc.PLAYER_MSG.ITEM_OBTAINED) {
+                        self.recivedItemRecord[data.id] ??= 0
+                        self.recivedItemRecord[data.id] += data.amount
+                    }
+                }
+            })()
+        )
+        sc.GameModel.inject({
+            enterTitle() {
+                this.parent()
+                lastArea = undefined
+                lastMapKeys = undefined
+            },
+        })
+
         sc.GameModel.inject({
             enterTitle() {
                 this.parent()
@@ -193,6 +215,7 @@ export class TextGather {
 
     private initGather() {
         let prevChar: string | undefined
+        const self = this
 
         let sideMsg: boolean = false
         sc.SideMessageHudGui.inject({
@@ -370,6 +393,26 @@ export class TextGather {
             onLevelUpEventEnd() {
                 this.parent()
                 interrupt()
+            },
+        })
+
+        sc.ItemHudBox.inject({
+            update() {
+                this.parent()
+                for (const id in self.recivedItemRecord) {
+                    let itemName = new sc.ItemContent(id).textGui.text!.toString().trim()
+                    if (itemName.endsWith('x')) itemName = itemName.slice(0, itemName.length - 1)
+                    const item = sc.inventory.items[parseInt(id)]
+                    const rarity = Lang.menu.itemRarityMap[Object.entries(sc.ITEMS_RARITY).find(e => e[1] == item.rarity)![0] as keyof typeof sc.ITEMS_RARITY]
+                    const type = Lang.menu.itemTypesMap[Object.entries(sc.ITEMS_TYPES).find(e => e[1] == item.type)![0] as keyof typeof sc.ITEMS_TYPES]
+                    const count = self.recivedItemRecord[id]
+                    if (count == 1) {
+                        speak(Lang.menu.oneItemRecivedTemplate.supplant({ itemName, rarity, type }))
+                    } else {
+                        speak(Lang.menu.multipleItemsRecivedTemplate.supplant({ itemName, count, rarity, type }))
+                    }
+                    delete self.recivedItemRecord[id]
+                }
             },
         })